# node-course/swagger/api.yaml
swagger: "2.0"
info:
  title: Demo API test
  description: This is my first swagger api demo
  version: "1.0"
  contact:
    email: arjun@gmail.com
  license:
    name: "MIT License"
    url: https://github.com/
  servers: 
    -url: http://localhost:4000/
    description: THis is my local server
    -url: http://localhost:3000
    description: This is my production server
schemes: [http, https]
paths: 
...
  /users/{id}:
    get: 
      summary: Return a object
      description: This will return the perticular user details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Success
          schema:
            type: object
            properties:
              id:
                type: integer
              name:
                type: string
        404: 
          description: Not Found

  # /create:
  #   post: 
  #     summary: Create post
  #     description: It will create a new post 
  #     producers:
  #       - application/json
  #     consumes:
  #       - application/json
  #     parameters:
  #       - in: body
  #         name: User body
  #         schema:
  #           type: object
  #           properties:
  #             id:
  #               type: integer
  #             name:
  #               type: string
  #     responses:
  #       200:
  #         description: Success
  #       201:
  #         description: created
...